ci: do dry runs of `publish-docs` for RCs
In several recent releases, the final `publish-docs` has failed to
complete, since it has no way of being tested beforehand. This adds a
version of the job that runs on RCs and does most of the same things but
stops short of the actual publish.

Ideally, we would have an even more complete dry run that did an actual
publish to some sort of staging environment, but that would be rather
more involved to set up and I believe this much of a dry run would
already have caught the recent failures.